fix: Remove dead code and simplify try/catch
